SwingSet: rewrite to new evaluate API, remove build-kernel step
The kernel is now loaded with bundle-source at launch time, removing the need
for a build-kernel step (and removing src/bundles/kernel).

`@agoric/evaluate` now provides evaluateProgram and evaluateBundle, and we
use the latter to load the kernel and static vat sources (in `nestedEvaluate`
mode, for better debugging metadata).

Dynamic vats are still loaded with evaluateProgram, for now. Soon this should
change to evaluateBundle for the same reasons.